refact: unify e2e checkButtonState custom command
